misc: rename stock shares folder
