+++ src/components/currentgame/stock-shares/StockShare.vue
