+++ src/components/currentgame/stock-shares/StockSharesBlock.vue
